+++ core/src/executive/executive.rs
@@ -176,7 +176,7 @@ impl<'a> CallCreateExecutive<'a> {
         } else if let Some(_) =
             internal_contract_map.contract(&params.code_address)
         {
-            info!(
+            debug!(
                 "CallInternalContract: address={:?} data={:?}",
                 params.code_address, params.data
             );
@@ -579,6 +579,7 @@ impl<'a> CallCreateExecutive<'a> {
                     } else {
                         Ok(())
                     };
+                    debug!("Internal Call Result: {:?}", result);
                     if let Err(e) = result {
                         state.revert_to_checkpoint();
                         Err(e.into())
@@ -747,7 +748,10 @@ impl<'a> CallCreateExecutive<'a> {
                 params,
                 mut unconfirmed_substate,
             ) => {
-                debug!("CallCreateExecutiveKind::ExecCreate");
+                debug!(
+                    "CallCreateExecutiveKind::ExecCreate: contract_addr = {:?}",
+                    params.address
+                );
                 assert!(self.is_create);
 
                 {