+++ core/src/executive/executive_tests.rs
@@ -48,7 +48,7 @@ fn test_contract_address() {
             CreateContractAddress::FromSenderNonceAndCodeHash,
             &address,
             &U256::from(88),
-            &[]
+            &[],
         )
         .0
     );
@@ -193,6 +193,71 @@ fn test_create_contract_out_of_depth() {
 
     assert_eq!(gas_left, U256::from(62_970));
     assert_eq!(substate.contracts_created.len(), 0);
+}
+
+#[test]
+fn test_suicide_when_creation() {
+    let factory = Factory::new(VMType::Interpreter, 1024 * 32);
+
+    // code:
+    //
+    // 33 - get caller address
+    // ff - self-deconstruct
+
+    let code = "33ff".from_hex().unwrap();
+
+    let sender_addr =
+        Address::from_str("1d1722f3947def4cf144679da39c4c32bdc35681").unwrap();
+    let contract_addr = contract_address(
+        CreateContractAddress::FromSenderNonceAndCodeHash,
+        &sender_addr,
+        &U256::zero(),
+        &[],
+    )
+    .0;
+
+    let mut params = ActionParams::default();
+    params.address = contract_addr;
+    params.sender = sender_addr;
+    params.original_sender = sender_addr;
+    params.storage_owner = sender_addr;
+    params.gas = U256::from(100_000);
+    params.code = Some(Arc::new(code));
+    params.value = ActionValue::Transfer(U256::from(0));
+
+    let storage_manager = new_state_manager_for_unit_test();
+    let mut state =
+        get_state_for_genesis_write_with_factory(&storage_manager, factory);
+    state
+        .add_balance(&sender_addr, &U256::from(100_000), CleanupMode::NoEmpty)
+        .unwrap();
+    let env = Env::default();
+    let machine = make_byzantium_machine(0);
+    let internal_contract_map = InternalContractMap::new();
+    let spec = machine.spec(env.number);
+    let mut substate = Substate::new();
+
+    let mut ex = Executive::new(
+        &mut state,
+        &env,
+        &machine,
+        &spec,
+        &internal_contract_map,
+    );
+    let FinalizationResult {
+        gas_left,
+        apply_state,
+        return_data: _,
+    } = ex.create(params, &mut substate).unwrap();
+
+    assert_eq!(gas_left, U256::from(94_998));
+    assert_eq!(apply_state, true);
+
+    assert!(substate.storage_collateralized.get(&sender_addr).is_none(),
+            "Since the contract has not been created, the sender occupied no storage now. ");
+    assert!(substate.storage_released.get(&sender_addr).is_none(),
+            "Since the contract has not been created, no storage is released when contract suicides. ");
+    assert!(substate.suicides.contains(&contract_addr));
 }
 
 #[test]
@@ -1449,7 +1514,7 @@ fn test_storage_commission_privilege() {
             .check_collateral_for_storage_finally(
                 &privilege_control_address,
                 &U256::MAX,
-                &mut substate
+                &mut substate,
             )
             .unwrap(),
         CollateralCheckResult::Valid
@@ -1749,7 +1814,7 @@ fn test_storage_commission_privilege() {
             .check_collateral_for_storage_finally(
                 &privilege_control_address,
                 &U256::MAX,
-                &mut substate
+                &mut substate,
             )
             .unwrap(),
         CollateralCheckResult::Valid